Complete testing on DEI endpoint
- Switch to DatabaseSession for object management
- Edit alembic revision with auto-increment
- Fix logic errors
