+++ augur/application/schema/alembic/versions/20_add_dei_badging_table.py
@@ -1,0 +1,40 @@
+"""Add DEI Badging table
+
+Revision ID: 20
+Revises: 19
+Create Date: 2023-06-15 06:02:42.082872
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '20'
+down_revision = '19'
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('dei_badging',
+    sa.Column('id', sa.Integer, nullable=False, autoincrement=True),
+    sa.Column('badging_id', sa.Integer(), nullable=False),
+    sa.Column('level', sa.String(), nullable=False),
+    sa.Column('repo_id', sa.BigInteger(), nullable=False),
+    sa.ForeignKeyConstraint(['repo_id'], ['augur_data.repo.repo_id'], name='user_repo_user_id_fkey'),
+    sa.PrimaryKeyConstraint('id', 'repo_id'),
+    schema='augur_data'
+    )
+    
+    op.add_column('users', sa.Column('email_verified', sa.Boolean(), server_default='false', nullable=False), schema='augur_operations')
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column('users', 'email_verified', schema='augur_operations')
+    
+    op.drop_table('dei_badging', schema='augur_data')
+    # ### end Alembic commands ###
